refactor: move server and tests into modules
